refactor createStore and part of the joins
